[alphaomega] Improve docker replacement with podman
Keeps docker enabled for now, due to build inconsistencies with
current work projects.

Adds wrapper derivation to symlink docker to podman on PATH. In
addition it removes the alias of docker to podman, because aliases are
not expanded in subshells.

Finally, refines the container policy to accept by default anything in
container-storage.